Lab1/requirements.txt
grpcio==1.60.1
protobuf==4.25.3
grpcio-tools==1.60.1